debian: update info about 64-bit toolchain
The info about 64-bit toolchain is a bit outdated.  For example, in my
testing, GCC 6.3 and 7.1 work for UEFI, and there is no compiling error
seen with bootloader and Linux kernel build.  Furthermore, the latest
U-Boot requires GCC higher than 6.0.  So let's remove the limit of GCC
version and recommend a resent GCC.

Signed-off-by: Shawn Guo <shawn.guo@linaro.org>