OP-TEE.md: add BL32_EXTRA1 and BL32_EXTRA2 for ATF fip image
When building optee_os in 32-bit, we will see following warning in ATF
boot log.

 WARNING: Not a known TEE, use default loading options.
 INFO:    Invalid OPTEE header, set legacy mode.

And ATF will just hang at the following point.

 INFO:    BL31: Initializing BL32

Although the same warning can be seen with 64-bit optee_os build, ATF
doesn't hang and optee works just fine.  The reason is that
tee-pager.bin image doesn't contain a TEE header, and ATF will always
try to boot it in 64-bit mode.  But for 32-bit build of optee_os, it
will fail.

To get the build work with both 32-bit and 64-bit optee_os, we need to
use v2 optee images to compose ATF fip image.  In that case, the
optee_os image will have a header, and ATF will check the header to
decide it's a 32-bit or 64-bit booting.

Signed-off-by: Shawn Guo <shawn.guo@linaro.org>